[refactor]extract common components to AuthInput
